% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructParents.R
\name{constructCustomerParents}
\alias{constructCustomerParents}
\title{Customer hierarchy}
\usage{
constructCustomerParents(customer)
}
\arguments{
\item{customer}{e.g. \code{all_tabs$customer}}
}
\value{
data dataframe of customer names and parents
}
\description{
Customer hierarchy
}
\examples{
all_tabs <- getPFData()
customerParents <- constructCustomerParents(all_tabs$customer)
}
