% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructParents.R
\name{constructProjectParents}
\alias{constructProjectParents}
\title{Project hierarchy}
\usage{
constructProjectParents(project)
}
\arguments{
\item{project}{e.g. \code{all_tabs$project}}
}
\value{
data dataframe of project names and parents
}
\description{
Project hierarchy
}
\examples{
all_tabs <- getPFData()
projectParents <- constructProjectParents(all_tabs$project)


# diagnostics
table(projectParents$ProjectName2, projectParents$ctu_division, useNA = "ifany")
x <- projectParents \%>\% 
  dplyr::group_by(ctu_projectName, ProjectName1) \%>\% 
  dplyr::summarize(n_divs = length(unique(ctu_division)))


}
