% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeCustomFields.R
\name{decodeCustomFields}
\alias{decodeCustomFields}
\title{Add custom fields to projects data}
\usage{
decodeCustomFields(project, customfields)
}
\arguments{
\item{project}{e.g. \code{all_tabs$project}}

\item{customfields}{e.g. \code{all_tabs$customfields}}
}
\description{
Add custom fields to projects data
}
\examples{
project <- all_tabs$project
customfields <- all_tabs$customfields
project2 <- decodeCustomFields(project, customfields)
}
