% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_all_NA.R
\name{discard_all_NA}
\alias{discard_all_NA}
\title{Discard dataframes without any data (all NA)}
\usage{
discard_all_NA(.x)
}
\arguments{
\item{.x}{dataframe}
}
\description{
Discard dataframes without any data (all NA)
}
\examples{
data.frame(a = 1, b = NA, c = 9) \%>\% 
  discard_all_NA()

}
