% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPFData.R
\name{downloadPFData}
\alias{downloadPFData}
\title{Download data from projectfacts via ODBC}
\usage{
downloadPFData(all = FALSE, ...)
}
\arguments{
\item{all}{logical - export all tables in the database (TRUE) or just the useful ones (FALSE)}
}
\description{
Download data from projectfacts via ODBC
}
\examples{
# downloadPFData()
}
