% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findvar.R
\name{findvar}
\alias{findvar}
\title{Find which dataframes contain a given variable}
\usage{
findvar(all_tabs, x)
}
\arguments{
\item{all_tabs}{e.g. all_tabs <- getPFData()}

\item{x}{variable name to look for (case is ignored)}
}
\description{
Find which dataframes contain a given variable
}
\examples{
all_tabs <- getPFData()
findvar(all_tabs, "parent")
}
