% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTime.R
\name{prepTime}
\alias{prepTime}
\title{prepare pf time entries}
\usage{
prepTime(all_tabs)
}
\arguments{
\item{all_tabs}{result from e.g. getPFData}
}
\value{
dataframe
}
\description{
the output of this can be grouped by whatever variable (ctu_division, worker, ...) 
to create summaries of minutes worked at the designated level
}
\examples{
all_tabs <- getPFData()
x <- prepTime(all_tabs)
}
