% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectsize.R
\name{projectSize}
\alias{projectSize}
\title{Estimate project size based on planned number of hours}
\usage{
projectSize(all_tabs)
}
\arguments{
\item{all_tabs}{list of tables (i.e. result of \code{getPFData()})}
}
\value{
dataframe with total time planned for a project and the size of the project.
}
\description{
Small projects are < 200 hours, medium are 200 - 500 hours, large are > 500 hours
}
