% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setEncoding.R
\name{setEncoding}
\alias{setEncoding}
\title{Set encoding of character variables to \code{latin-1}}
\usage{
setEncoding(tabs)
}
\arguments{
\item{tabs}{e.g. all_tabs}
}
\description{
Set encoding of character variables to \code{latin-1}
}
\examples{
all_tabs <- getPFData() \%>\% setEncoding()
}
