% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pf.R
\name{update_pf}
\alias{update_pf}
\title{Update the installed version of the pf package}
\usage{
update_pf(...)
}
\arguments{
\item{...}{additional arguments to \code{link[remotes]{install_local}}}
}
\description{
Update the installed version of the pf package
}
\examples{
pf::update_pf()
}
