% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerActiveList.R
\name{workerActiveList}
\alias{workerActiveList}
\title{List of workers with their active status}
\usage{
workerActiveList(all_tabs)
}
\arguments{
\item{all_tabs}{e.g. the result from \code{getPFData()}}
}
\description{
List of workers with their active status
}
\examples{
getPFData() \%>\% 
setEncoding() \%>\% 
workerActiveList()
}
